+++ elmyra.ip.access.epo/CHANGES.rst
@@ -4,6 +4,7 @@ elmyra.ip.access.epo CHANGES
 
 development
 ===========
+- middleware: number normalization: AT362828E should be returned as AT362828T for querying at OPS
 
 0.84.0
 ======